third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-013-expected.txt


FAIL CSS Test: @viewport constrained - max-zoom not affecting width assert_equals: expected 320 but got 640
